# packages/contracts/openapi/orgawork.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OrgaWork API",
    "version": "1.0.0",
    "description": "قرارداد پایه و نسخه‌دار رابط برنامه‌نویسی OrgaWork"
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:3001",
      "description": "محیط محلی"
    }
  ],
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "operationId": "login",
        "summary": "ورود امن",
        "responses": {
          "200": {
            "description": "عملیات موفق بود",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "درخواست معتبر نیست",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "logout",
        "summary": "خروج",
        "responses": {
          "200": {
            "description": "عملیات موفق بود",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "درخواست معتبر نیست",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "نشست معتبر لازم است",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/logout-all": {
      "post": {
        "operationId": "logoutAll",
        "summary": "خروج از همه نشست‌ها",
        "responses": {
          "200": {
            "description": "عملیات موفق بود",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "درخواست معتبر نیست",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "نشست معتبر لازم است",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/session": {
      "get": {
        "operationId": "getSession",
        "summary": "مشاهده نشست جاری",
        "responses": {
          "200": {
            "description": "عملیات موفق بود",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "درخواست معتبر نیست",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "نشست معتبر لازم است",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/password-reset/request": {
      "post": {
        "operationId": "requestPasswordReset",
        "summary": "درخواست بازیابی گذرواژه",
        "responses": {
          "200": {
            "description": "عملیات موفق بود",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "درخواست معتبر نیست",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/password-reset/confirm": {
      "post": {
        "operationId": "confirmPasswordReset",
        "summary": "تأیید بازیابی گذرواژه",
        "responses": {
          "200": {
            "description": "عملیات موفق بود",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "درخواست معتبر نیست",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/organizations": {
      "get": {
        "operationId": "listOrganizations",
        "summary": "فهرست سازمان‌های فعال",
        "responses": {
          "200": {
            "description": "عملیات موفق بود",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "درخواست معتبر نیست",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "نشست معتبر لازم است",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/current-organization": {
      "post": {
        "operationId": "switchCurrentOrganization",
        "summary": "تغییر سازمان جاری",
        "responses": {
          "200": {
            "description": "عملیات موفق بود",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "description": "درخواست معتبر نیست",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "نشست معتبر لازم است",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "schemas": {
      "Uuid": {
        "type": "string",
        "format": "uuid"
      },
      "UtcTimestamp": {
        "type": "string",
        "format": "date-time"
      },
...
      "ResponseMeta": {
        "type": "object",
        "additionalProperties": false,
        "required": ["requestId", "correlationId", "timestamp", "contractVersion"],
        "properties": {
          "requestId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "correlationId": {
            "$ref": "#/components/schemas/Uuid"
          },
          "timestamp": {
            "$ref": "#/components/schemas/UtcTimestamp"
          },
          "contractVersion": {
            "const": "1.0.0"
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["ok", "data", "meta"],
        "properties": {
          "ok": {
            "const": true
          },
          "data": {},
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["ok", "error", "meta"],
        "properties": {
          "ok": {
            "const": false
          },
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_ERROR",
                  "UNAUTHENTICATED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "RATE_LIMITED",
                  "SERVICE_UNAVAILABLE",
                  "INTERNAL_ERROR",
                  "AUTH_INVALID_CREDENTIALS",
                  "AUTH_ACCOUNT_DISABLED",
                  "AUTH_SESSION_REQUIRED",
                  "AUTH_SESSION_EXPIRED",
                  "AUTH_SESSION_REVOKED",
                  "AUTH_CSRF_INVALID",
                  "AUTH_PASSWORD_POLICY_FAILED",
                  "AUTH_PASSWORD_RESET_INVALID",
                  "ORGANIZATION_CONTEXT_REQUIRED",
                  "ORGANIZATION_MEMBERSHIP_REQUIRED",
                  "ORGANIZATION_MEMBERSHIP_INACTIVE",
                  "ORGANIZATION_SWITCH_FORBIDDEN",
                  "AUTHORIZATION_DENIED",
                  "INVITATION_INVALID",
                  "INVITATION_EXPIRED",
                  "INVITATION_REVOKED",
                  "MEMBERSHIP_STATE_CONFLICT",
                  "TEAM_ORGANIZATION_MISMATCH"
                ]
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "field": {
                "type": "string"
              }
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        }
      },
...
    }
  }
}